samples/python/agents/a2a_mcp/pyproject.toml
[project]
name = "a2a-mcp"
version = "0.1.0"
description = "A2A - MCP Sample"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[sql]",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "google-adk>=1.0.0",
    "httpx>=0.28.1",
    # Added for DashScope / OpenAI compatible integration
    "langchain-openai>=0.2.5",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.1",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "litellm",
    "openai>=1.97.1",
    # Added for DashScope embedding via llama-index per user request
    "llama-index>=0.11.0",
    "llama-index-embeddings-dashscope>=0.2.0",
    "dashscope>=1.22.3",
]

# Optional dependency groups (uv will treat these as PEP 735 groups).
[dependency-groups]
dev = [
    "debugpy>=1.8.5",
]

[project.scripts]
a2a-mcp = "a2a_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
